First Commit of pytorch * Added PyTorch Demo Jupyter Notebook * Rest is hodgepodge for now
