+++ src/models_tf/base_infergan.py
